--- js/avatar.js
@@ -1,10 +1,0 @@
-import {getRandomNumber} from './random';
-
-const getAvatar = () => {
-  const avatarId = getRandomNumber(1, 9);
-  const author = {
-    avatar: `img/avatars/user0${avatarId}.png`};
-  return author;
-};
-
-export {getAvatar};
